refactor(server)!: read auth from ctx headers in review_document
- Replace mcpo_headers with required ctx and read Authorization
  from ctx.request_context.request.headers; fall back to admin
  TOKEN if missing
- Pass token explicitly to download_file
- Expand review_document docstring with index rules for DOCX
  (pid:), XLSX (A1), PPTX (sid:) and note markdown download link
- Add sizing unit imports (docx.shared.Inches, pptx.util.Inches,
  Pt as PptPt)
- Clean up unused comments and noqa

Why: unify and harden auth handling via request context, improve API clarity, and prepare for consistent sizing utilities.

BREAKING CHANGE: review_document signature changed; mcpo_headers removed and ctx is now required.
